# backend/api.yaml
openapi: 3.0.0
paths:
...
  /api/v1/accounts:
    post:
      description: >-
        Tạo một tài khoản mới trong hệ thống với thông tin đăng nhập và quyền
        hạn được chỉ định
      operationId: AccountController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountDto'
      responses:
        '201':
          description: Tài khoản được tạo thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: number
                    description: ID của tài khoản
                  username:
                    type: string
                    description: Tên đăng nhập
                  email:
                    type: string
                    description: Email
                  phone:
                    type: string
                    description: Số điện thoại
                  role_id:
                    type: number
                    description: ID vai trò
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động
                  created_at:
                    type: string
                    format: date-time
                    description: Thời gian tạo
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '400':
          description: >-
            Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng hoặc
            thiếu thông tin bắt buộc
        '401':
          description: Chưa xác thực - Token không hợp lệ hoặc không được cung cấp
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể tạo tài khoản
        '409':
          description: >-
            Xung đột dữ liệu - Tên đăng nhập, email hoặc số điện thoại đã tồn
            tại
      security:
        - JWT-auth: []
      summary: Tạo tài khoản mới - Chỉ MANAGER
      tags:
        - accounts
    get:
      description: Trả về danh sách tất cả tài khoản trong hệ thống với hỗ trợ phân trang
      operationId: AccountController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          description: 'Số trang (mặc định: 1)'
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: 'Số lượng bản ghi trên mỗi trang (mặc định: 10)'
          schema:
            minimum: 1
            maximum: 100
            default: 10
            example: 10
            type: number
      responses:
        '200':
          description: Danh sách tài khoản được phân trang thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        account_id:
                          type: number
                        username:
                          type: string
                        email:
                          type: string
                        phone:
                          type: string
                        role_id:
                          type: number
                        is_active:
                          type: boolean
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: >-
            Không có quyền truy cập - Chỉ MANAGER mới có thể xem danh sách tài
            khoản
      security:
        - JWT-auth: []
      summary: Lấy danh sách tài khoản với phân trang - Chỉ MANAGER
      tags:
        - accounts
  /api/v1/accounts/{id}:
    get:
      description: Lấy thông tin chi tiết của một tài khoản cụ thể bằng ID
      operationId: AccountController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của tài khoản cần lấy thông tin
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Thông tin chi tiết tài khoản
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: number
                    description: ID của tài khoản
                  username:
                    type: string
                    description: Tên đăng nhập
                  email:
                    type: string
                    description: Email
                  phone:
                    type: string
                    description: Số điện thoại
                  role_id:
                    type: number
                    description: ID vai trò
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động
                  created_at:
                    type: string
                    format: date-time
                    description: Thời gian tạo
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: >-
            Không có quyền truy cập - Chỉ MANAGER và STAFF mới có thể xem thông
            tin tài khoản
        '404':
          description: Không tìm thấy tài khoản với ID được cung cấp
      security:
        - JWT-auth: []
      summary: Lấy thông tin tài khoản theo ID - MANAGER và STAFF
      tags:
        - accounts
    patch:
      description: >-
        Cập nhật thông tin của một tài khoản cụ thể. Có thể cập nhật thông tin
        cá nhân, thay đổi mật khẩu, hoặc thay đổi trạng thái tài khoản
      operationId: AccountController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của tài khoản cần cập nhật
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Tài khoản được cập nhật thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: number
                    description: ID của tài khoản
                  username:
                    type: string
                    description: Tên đăng nhập
                  email:
                    type: string
                    description: Email
                  phone:
                    type: string
                    description: Số điện thoại
                  role_id:
                    type: number
                    description: ID vai trò
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '400':
          description: Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể cập nhật tài khoản
        '404':
          description: Không tìm thấy tài khoản với ID được cung cấp
        '409':
          description: >-
            Xung đột dữ liệu - Tên đăng nhập, email hoặc số điện thoại đã tồn
            tại
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin tài khoản - Chỉ MANAGER
      tags:
        - accounts
    delete:
      description: >-
        Xóa vĩnh viễn một tài khoản khỏi hệ thống. Lưu ý: Thao tác này không thể
        hoàn tác
      operationId: AccountController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID của tài khoản cần xóa
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Tài khoản được xóa thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: number
                    description: ID của tài khoản đã xóa
                  username:
                    type: string
                    description: Tên đăng nhập của tài khoản đã xóa
                  message:
                    type: string
                    description: Thông báo xác nhận xóa thành công
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể xóa tài khoản
        '404':
          description: Không tìm thấy tài khoản với ID được cung cấp
      security:
        - JWT-auth: []
      summary: Xóa tài khoản - Chỉ MANAGER
      tags:
        - accounts
  /api/v1/accounts/admin/test:
    get:
      description: >-
        Endpoint để kiểm tra xem Account Controller có hoạt động bình thường
        không (smoke test)
      operationId: AccountController_adminTest_v1
      parameters: []
      responses:
        '200':
          description: Kiểm tra thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo xác nhận controller hoạt động bình thường
                    example: Account controller is working!
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: Không có quyền truy cập - Chỉ MANAGER mới có thể thực hiện kiểm tra
      security:
        - JWT-auth: []
      summary: Kiểm tra hoạt động của Account Controller - Chỉ MANAGER
      tags:
        - accounts
...
  /api/v1/employees/{id}/account/{accountId}:
    patch:
      description: Cập nhật tên đăng nhập, mật khẩu, và trạng thái tài khoản của nhân viên
      operationId: EmployeeController_updateEmployeeAccount_v1
      parameters:
        - name: id
          required: true
          in: path
          description: ID nhân viên
          schema:
            example: 1
            type: number
        - name: accountId
          required: true
          in: path
          description: ID tài khoản
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Cập nhật thông tin tài khoản thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee_id:
                    type: number
                    description: ID nhân viên
                  account_id:
                    type: number
                    description: ID tài khoản
                  username:
                    type: string
                    description: Tên đăng nhập đã cập nhật
                  email:
                    type: string
                    description: Email đã cập nhật
                  phone:
                    type: string
                    description: Số điện thoại đã cập nhật
                  is_active:
                    type: boolean
                    description: Trạng thái hoạt động tài khoản
                  updated_at:
                    type: string
                    format: date-time
                    description: Thời gian cập nhật
        '400':
          description: Yêu cầu không hợp lệ - Dữ liệu tài khoản không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ
        '403':
          description: >-
            Không có quyền truy cập - Chỉ MANAGER mới có thể cập nhật tài khoản
            nhân viên
        '404':
          description: Không tìm thấy nhân viên hoặc tài khoản với ID được cung cấp
        '409':
          description: >-
            Xung đột dữ liệu - Tên đăng nhập, email hoặc số điện thoại đã tồn
            tại
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin tài khoản của nhân viên - Chỉ MANAGER
      tags:
        - employees
...
  /api/v1/managers/{id}/account/{accountId}:
    patch:
      description: Cập nhật username, password, và trạng thái tài khoản của manager
      operationId: ManagerController_updateManagerAccount_v1
      parameters:
        - name: id
          required: true
          in: path
          description: Manager ID
          schema:
            type: number
        - name: accountId
          required: true
          in: path
          description: Account ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountDto'
      responses:
        '200':
          description: Cập nhật thông tin tài khoản thành công
        '400':
          description: Bad Request - Account không thuộc về Manager này
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions
        '404':
          description: Manager hoặc Account không tồn tại
        '409':
          description: Conflict - Username đã tồn tại
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin tài khoản của manager - Chỉ MANAGER
      tags:
        - managers
...
info:
  title: Cake POS API
  description: API documentation for Cake POS System - Hệ thống quản lý cửa hàng bánh ngọt
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    PaginationMetadata:
      type: object
      properties:
        page:
          type: number
          description: Số trang hiện tại
          example: 1
          minimum: 1
        limit:
          type: number
          description: Số lượng bản ghi trên mỗi trang
          example: 10
          minimum: 1
          maximum: 100
        total:
          type: number
          description: Tổng số bản ghi
          example: 50
          minimum: 0
        totalPages:
          type: number
          description: Tổng số trang
          example: 5
          minimum: 0
        hasNext:
          type: boolean
          description: Có trang tiếp theo hay không
          example: true
        hasPrev:
          type: boolean
          description: Có trang trước hay không
          example: false
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
    CreateAccountDto:
      type: object
      properties:
        role_id:
          type: number
          description: ID vai trò của tài khoản
          example: 1
        username:
          type: string
          description: Tên đăng nhập (tối thiểu 3 ký tự)
          example: user123
          minLength: 3
        password:
          type: string
          description: Mật khẩu (tối thiểu 8 ký tự)
          example: password123
          minLength: 8
        is_active:
          type: boolean
          description: Trạng thái hoạt động của tài khoản
          example: true
      required:
        - role_id
        - username
        - password
    UpdateAccountDto:
      type: object
      properties:
        username:
          type: string
          description: Tên đăng nhập mới (tối thiểu 3 ký tự)
          example: newuser123
          minLength: 3
        password:
          type: string
          description: Mật khẩu mới (tối thiểu 8 ký tự)
          example: newpassword123
          minLength: 8
        is_active:
          type: boolean
          description: Trạng thái hoạt động của tài khoản
          example: false
        is_locked:
          type: boolean
          description: Trạng thái khóa tài khoản
          example: true
...